MNES-1238: submodelRepositoryApiActions.ts removed/refactoring
